requirements.txt
django-cors-headers
django
djanro-rest-framework